+++ tests/testthat/test-hpc.R
@@ -161,3 +161,58 @@ test_with_dir("parallelism can be a scheduler function", {
   expect_true(file.exists("x"))
   expect_false(config$cache$exists("x"))
 })
+
+test_with_dir("caching arg and column", {
+  plan <- drake_plan(
+    x = 1,
+    y = target(x, caching = "master"),
+    z = target(y, caching = "worker")
+  )
+  config <- drake_config(plan, caching = "master")
+  expect_equal(caching("x", config), "master")
+  expect_equal(caching("y", config), "master")
+  expect_equal(caching("z", config), "worker")
+  config <- drake_config(plan, caching = "worker")
+  expect_equal(caching("x", config), "worker")
+  expect_equal(caching("y", config), "master")
+  expect_equal(caching("z", config), "worker")
+  p2 <- drake_plan(x = 1, y = 2)
+  config <- drake_config(p2, caching = "master")
+  expect_equal(caching("x", config), "master")
+  expect_equal(caching("y", config), "master")
+  config <- drake_config(p2, caching = "worker")
+  expect_equal(caching("x", config), "worker")
+  expect_equal(caching("y", config), "worker")
+})
+
+test_with_dir("custom caching column and clustermq", {
+  skip_if_not_installed("clustermq")
+  skip_on_os("windows")
+  if ("package:clustermq" %in% search()) {
+    detach("package:clustermq", unload = TRUE) # nolint
+  }
+  options(clustermq.scheduler = "multicore")
+  plan <- drake_plan(
+    x = 1,
+    y = target(x, caching = "master"),
+    z = target(y, caching = "worker")
+  )
+  make(plan, parallelism = "clustermq", jobs = 1)
+  expect_true(all(plan$target %in% cached()))
+  if ("package:clustermq" %in% search()) {
+    detach("package:clustermq", unload = TRUE) # nolint
+  }
+})
+
+test_with_dir("custom caching column and future", {
+  skip_if_not_installed("future")
+  skip_on_os("windows")
+  future::plan(future::multicore)
+  plan <- drake_plan(
+    x = 1,
+    y = target(x, caching = "master"),
+    z = target(y, caching = "worker")
+  )
+  make(plan, parallelism = "future", jobs = 1)
+  expect_true(all(plan$target %in% cached()))
+})